requirements.txt
rq
redis
loguru
python-dotenv
